pyproject.toml
[project]
name = "graduate-info"
version = "0.1.0"
description = "研究生调剂信息爬虫"
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
dependencies = [
    "scrapy>=2.11.0",
    "pymysql>=1.1.0",
    "sqlalchemy>=2.0.0",
    "python-dotenv>=1.0.0",
    "mysql-connector-python>=8.0.0",
    "requests>=2.32.3",
    "autopep8>=2.3.2",
]
requires-python = ">=3.9"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 120
target-version = "py39"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
